requirements.txt
google-adk
dotenv
sqlalchemy
nest_asyncio
python-docx 
pandas 
markdown 
pypdf